feat: further split form into drag and drop
