+++ iqps-frontend/src/components/Upload/UploadForm.tsx
@@ -1,10 +1,10 @@
-import { ChangeEvent, createRef, Dispatch, DragEventHandler, MouseEventHandler, useState } from "react";
+import { createRef, Dispatch, MouseEventHandler, useState } from "react";
 import { IQuestionPaperFile } from "../../types/question_paper";
 import { FileCard } from "./FileCard";
 import Spinner from "../Spinner/Spinner";
 import { AiOutlineCloudUpload, AiOutlineFileAdd } from "react-icons/ai";
 import { validate } from "../../utils/validateInput";
-import toast from "react-hot-toast";
+import { UploadDragAndDrop } from "./UploadDragAndDrop";
 
 interface IUploadFormProps {
 	max_upload_limit: number;
@@ -15,43 +15,11 @@ interface IUploadFormProps {
 }
 export function UploadForm(props: IUploadFormProps) {
 	const [qPapers, setQPapers] = useState<IQuestionPaperFile[]>([]);
-	const [isDragging, setIsDragging] = useState<boolean>(false);
 	const [selectedQPaper, setSelectedQPaper] =
 		useState<IQuestionPaperFile | null>(null);
 
-	const fileInputRef = createRef<HTMLInputElement>();
-
-	const onDragEnter: DragEventHandler<HTMLDivElement> = (e) => {
-		e.preventDefault();
-		e.stopPropagation();
-		setIsDragging(true);
-	}
-
-	const onDragExit: DragEventHandler<HTMLDivElement> = (e) => {
-		e.preventDefault();
-		e.stopPropagation();
-		setIsDragging(false);
-	}
-
-	const openFileDialog: MouseEventHandler = (e) => {
-		e.stopPropagation();
-		fileInputRef.current?.click();
-	};
-
 	const isQPValid = (data: IQuestionPaperFile) => {
 		return !Object.values(validate(data)).some(Boolean);
-	};
-
-	const onFileInputChange = async (e: ChangeEvent<HTMLInputElement>) => {
-		e.preventDefault();
-		if (e.target) {
-			const newFiles = Array.from(
-				(e.target as HTMLInputElement).files || []
-			);
-			if (newFiles) {
-				await addQPapers(newFiles);
-			}
-		}
 	};
 
 	const addQPapers = async (newFiles: File[]) => {
@@ -80,87 +48,53 @@ export function UploadForm(props: IUploadFormProps) {
 		);
 	};
 
-	const onFileDrop: DragEventHandler<HTMLDivElement> = async (e) => {
-		e.preventDefault();
-
-		if (e.dataTransfer) {
-			const pdfFiles = [...e.dataTransfer.files].filter(
-				(file) => file.type === "application/pdf"
-			);
-			if (pdfFiles && pdfFiles.length > 0) {
-				if (pdfFiles.length > props.max_upload_limit) {
-					toast.error(`max ${props.max_upload_limit} files allowed`);
-					return;
-				}
-				await addQPapers(pdfFiles);
-			} else {
-				toast.error("Could not catch files. Please try again");
-			}
-			e.dataTransfer.clearData();
-		}
-
-		setIsDragging(false);
+	const fileInputRef = createRef<HTMLInputElement>();
+	const openFileDialog: MouseEventHandler = (e) => {
+		e.stopPropagation();
+		fileInputRef.current?.click();
 	};
 
 	return <div className="upload-section">
-		{qPapers.length > 0 ? (
-			<>
-				<div className="uploaded-files">
-					{qPapers.map(
-						(qp, i) => <div key={i}>
-							<FileCard
-								qPaper={qp}
-								removeQPaper={removeQPaper}
-								edit={setSelectedQPaper}
-							/>
-							{!isQPValid(qp) && (
-								<p className="error-msg">
-									Invalid course details
-								</p>
-							)}
-						</div>
-					)}
-				</div>
-				<div className="upload-section-btns">
-					<button onClick={props.handleUpload} className="upload-btn">
-						{props.awaitingResponse ? (
-							<>
-								Uploading
-								<div className="spinner">
-									<Spinner />
-								</div>
-							</>
-						) : (
-							<><AiOutlineCloudUpload size="1.5rem" />Upload</>
+		{
+			qPapers.length > 0 ? (
+				<>
+					<div className="uploaded-files">
+						{qPapers.map(
+							(qp, i) => <div key={i}>
+								<FileCard
+									qPaper={qp}
+									removeQPaper={removeQPaper}
+									edit={setSelectedQPaper}
+								/>
+								{!isQPValid(qp) && (
+									<p className="error-msg">
+										Invalid course details
+									</p>
+								)}
+							</div>
 						)}
-					</button>
-					{qPapers.length <= props.max_upload_limit && <button onClick={openFileDialog}>
-						<AiOutlineFileAdd size="1.5rem" />Add More Files
-					</button>}
-				</div>
-			</>
-		) : (
-			!props.awaitingResponse && (
-				<div
-					className={`upload-area ${isDragging && "active"}`}
-					onDragOver={onDragEnter}
-					onDragLeave={onDragExit}
-					onDrop={onFileDrop}
-					onClick={openFileDialog}
-				>
-					<input
-						ref={fileInputRef}
-						type="file"
-						accept=".pdf"
-						hidden
-						multiple={true}
-						onChange={onFileInputChange}
-					/>
-					<AiOutlineCloudUpload className="upload-icon" size="5rem" />
-					<h2>Click or drop files to upload</h2>
-				</div>
+					</div>
+					<div className="upload-section-btns">
+						<button onClick={props.handleUpload} className="upload-btn">
+							{props.awaitingResponse ? (
+								<>
+									Uploading
+									<div className="spinner">
+										<Spinner />
+									</div>
+								</>
+							) : (
+								<><AiOutlineCloudUpload size="1.5rem" />Upload</>
+							)}
+						</button>
+						{qPapers.length <= props.max_upload_limit && <button onClick={openFileDialog}>
+							<AiOutlineFileAdd size="1.5rem" />Add More Files
+						</button>}
+					</div>
+				</>
+			) : (
+				!props.awaitingResponse && <UploadDragAndDrop max_upload_limit={props.max_upload_limit} fileInputRef={fileInputRef} addQPapers={addQPapers} openFileDialog={openFileDialog} />
 			)
-		)
 		}
 	</div >;
 }
